% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rpublons-package}
\alias{rpublons-package}
\alias{rpublons}
\alias{publons}
\alias{publons_auth}
\title{rpublons}
\usage{
publons(verb = c("GET", "POST", "HEAD", "OPTIONS"), endpoint, body = NULL,
  query = NULL, ..., base_url = "https://publons.com/api/v2/",
  token = Sys.getenv("PUBLONS_TOKEN"),
  username = Sys.getenv("PUBLONS_USERNAME"),
  password = Sys.getenv("PUBLONS_PASSWORD"))

publons_auth(username, password)
}
\arguments{
\item{verb}{A character string containing an HTTP request verb.}

\item{endpoint}{A character string containing an API endpoint.}

\item{body}{Optionally, a request body.}

\item{query}{Optionally, a list containing HTTP query arguments.}

\item{base_url}{A character string specifying the base URL for the API.}

\item{token}{A character string containing a Publons API token. By default, taken from environment variables.}

\item{username}{A character string, containing a Publons username, passed to \code{publons_auth} if \code{token} is missing. This is used to generate an API token, if needed.}

\item{password}{A character string, containing a Publons password, passed to \code{publons_auth} if \code{token} is missing. This is used to generate an API token, if needed.}
}
\description{
Client for Publons.com
}
\details{
Extract review information information from Publons \url{https://publons.com}. This is currently a low-level client that handles authorization but no high-level functionality.
}
\examples{
\dontrun{
  # retrieve and store API token
  publons_auth("myusername", "mypassword")
  
  # get list of academics
  publons("GET", "academic/")
  
  # get academics by institution
  publons("GET", "academic/", query = list(institution = "Harvard University"))
  
  # get a specific academic by ORCiD
  publons("GET", paste0("academic/", "0000-0003-4097-6326"))
  
  # get reviews for a specific academic by ORCiD
  publons("GET", "academic/review/", 
          query = list(academic = "0000-0003-4097-6326"))
  
  # get a specific review
  publons("GET", paste0("academic/review/", "YK8zafgo"))
  
}
}
\references{
\href{https://publons.com/api/v2/}{Publons API v2 Documentation}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
